modified:   include/meshtk/MeshViewer.hh 	modified:   include/meshtk/TriMesh.hh 	modified:   include/meshtk/mesh_assist.hh 	modified:   src/main.cc 	modified:   src/meshtk/DynamicTriMesh.cc 	modified:   src/meshtk/MeshViewer.cc 	modified:   src/meshtk/TriMesh_Base.cc 	modified:   src/meshtk/TriMesh_Curv.cc 	modified:   src/meshtk/TriMesh_SIFT.cc 	modified:   src/meshtk/mesh_assist.cc 	modified:   src/test.cc

+++ include/meshtk/MeshViewer.hh
@@ -33,13 +33,14 @@ namespace meshtk {
 
   protected:
     TriMesh *obj;
+
     GLuint vn; //vertex number
     GLuint fn; //facet number
   
 
 
-    GLfloat *vertex_array;
-    GLfloat *normal_array;
+    GLdouble *vertex_array;
+    GLdouble *normal_array;
     GLuint *index_array;
 
   public:
